Some progress getting tests to pass.

+++ include/data/math/number/gmp/Z.hpp
@@ -374,7 +374,7 @@ namespace data::encoding::hexidecimal {
     
 }
 
-namespace data::encoding::decimal {
+namespace data::encoding::integer {
     
     std::string write(const math::number::gmp::Z& n);
